fix: outbound flow (#2)
# Overview

There was an issue where a successful form submission would render a
page that stated a translation was missing. This was due to the redirect
going to the wrong path. It was pointing at the root path when I was
building the proof of concept, but it has since changed the target path.

This also makes it so that all companies in the database get emailed
rather than just California data brokers.