fix: form errors (#16)
# Overview

This fixes the alerts logic in `RequesterFormComponent` since there was
a bug with rendering a single error on the page. Before, we were only
checking for multiple errors before rendering an unordered list. Now,
we'll render a list of errors, a paragraph for a non-array error, and a
regular paragraph for a notice.