Sanitize company data more vigorously
This was partly due to some failing integration tests that I noticed
with today's CPPA data. There was a new and incorrectly formatted
website in the shape of `https//websitedomain`, so this adds more
whitespace extraction and a little bit of regex to vigorously clean data
before putting it into the database.